chore: extracted `TypeMismatchError` to separate file
