+++ packages/groqd/src/commands/project.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, expectTypeOf, it } from "vitest";
 import { SanitySchema, SchemaConfig } from "../tests/schemas/nextjs-sanity-fe";
 import { InferResultItem, InferResultType } from "../types/public-types";
-import { Simplify, TypeMismatchError } from "../types/utils";
+import { Simplify } from "../types/utils";
+import { TypeMismatchError } from "../types/type-mismatch-error";
 import { createGroqBuilderWithZod } from "../index";
 import { mock } from "../tests/mocks/nextjs-sanity-fe-mocks";
 import { executeBuilder } from "../tests/mocks/executeQuery";
