+++ packages/groqd/src/types/utils.ts
@@ -1,4 +1,5 @@
-import type { IsNever, Simplify } from "type-fest";
+import type { Simplify } from "type-fest";
+import { TypeMismatchError } from "./type-mismatch-error";
 
 export type { Simplify, Primitive, LiteralUnion, IsAny } from "type-fest";
 
@@ -10,8 +11,8 @@ export type Get<TObj, TKey> = TKey extends keyof TObj
   ? TObj[TKey]
   : TypeMismatchError<{
       error: "Invalid property";
-      expected: keyof TObj;
       actual: TKey;
+      expected: keyof TObj;
     }>;
 
 /**
@@ -31,31 +32,6 @@ export type SimplifyDeep<T> = T extends object
   : T;
 
 export type Override<T, TOverrides> = Omit<T, keyof TOverrides> & TOverrides;
-
-export type TypeMismatchError<
-  TError extends { error: string; expected: any; actual: any } = any
-> = {
-  error: TError["error"];
-  expected: Simplify<TError["expected"]>;
-  actual: Simplify<TError["actual"]>;
-};
-
-/**
- * Extracts all TypeMismatchError's from the projection result,
- * making it easy to report these errors.
- * Returns a string of error messages,
- * or `never` if there are no errors.
- */
-export type ExtractTypeMismatchErrors<TProjectionResult> = ValueOf<{
-  [TKey in StringKeys<
-    keyof TProjectionResult
-  >]: TypeMismatchError extends TProjectionResult[TKey]
-    ? `Error in "${TKey}": ${Extract<
-        TProjectionResult[TKey],
-        TypeMismatchError
-      >["error"]}`
-    : never;
-}>;
 
 /**
  * Returns a union of all value types in the object
@@ -133,20 +109,3 @@ export function pick<T, TKeys extends keyof T>(
 export type UndefinedToNull<T> = T extends undefined
   ? NonNullable<T> | null
   : T;
-/**
- * When we map projection results, we return TypeMismatchError's
- * for any fields that have an invalid mapping configuration.
- * However, this does not cause TypeScript to throw any errors.
- *
- * In order to get TypeScript to complain about these invalid mappings,
- * we will "require" an extra parameter, which will reveal the error messages.
- */
-export type RequireAFakeParameterIfThereAreTypeMismatchErrors<
-  TProjectionResult,
-  _Errors extends string = ExtractTypeMismatchErrors<TProjectionResult>
-> = IsNever<_Errors> extends true
-  ? [] // No errors, yay! Do not require any extra parameters.
-  : // We've got errors; let's require an extra parameter, with the error message:
-    | [_Errors]
-      // And this extra error message causes TypeScript to always log the entire list of errors:
-      | ["⛔️ Error: this projection has type mismatches: ⛔️"];
