+++ packages/groqd/src/validation/zod.test.ts
@@ -3,7 +3,7 @@ import { createGroqBuilderWithZod, InferResultType } from "../index";
 import { SchemaConfig } from "../tests/schemas/nextjs-sanity-fe";
 import { mock } from "../tests/mocks/nextjs-sanity-fe-mocks";
 import { executeBuilder } from "../tests/mocks/executeQuery";
-import { TypeMismatchError } from "../types/utils";
+import { TypeMismatchError } from "../types/type-mismatch-error";
 
 const q = createGroqBuilderWithZod<SchemaConfig>();
